Created the DataBase connectivity, as well as when you click on the submit button on the main WinForm it will open the customer DashBoard, please review and let me know what I should change  (#31)
* Fixing stuff

* I added the database connectivity, when you click on submit button it opens the customers dashboard, created Car Table and Reservation Table inside of the main method still need to double check if the db file is good

* Updated the login method

When the user clicks the Submit button, the program takes the ID and password entered, then checks if they match a record in the database. If the credentials are valid, it opens the CustomerDashboard; if not, the user sees an error message indicating an incorrect ID or password. This setup helps me ensure that only authorized users can access the dashboard.

* added CreateUserTable, InsertUsers, DB file now contains tables: Cars, Reservations, and Users

added CreateUserTable, InsertUsers,
DB file now contains tables: Cars, Reservations, and Users

Also tweaked your submit button, to use SQLite, but you should review your code

App to see: (DB Browser SQLite) note for myself

* 1 error left and Admin Class to do

*Admin class still has to be finished.
*There is still an error whenever you the submit button in the login screen.

---------

Co-authored-by: Matthew Macri <123660194+MatthewMacri@users.noreply.github.com>